fix: 安装路径更新及说明
