+++ README.md
@@ -3,8 +3,11 @@
 ## 项目背景
 由于达梦数据库官方提供的 `sqlalchemy2.0` 在实际应用中存在多种问题，且官方更新较慢，无法及时响应社区反馈。本项目旨在解决 `dm-sqlalchemy` 在 SQLAlchemy 2.0 适配中的错误和兼容性问题，为开发者提供更稳定的解决方案。
 
-## 当前使用达梦sqlalchemy版本
+## 达梦sqlalchemy版本
 - 2.0.1
+
+## 源码安装
+- pip install .
 
 ## 已修复问题
 1. **自增id返回异常(id值不对)问题**
